+++ src/net/dungeonrealms/listeners/BankListener.java
@@ -222,7 +222,7 @@ public class BankListener implements Listener {
                         } else if (e.isRightClick()) {
                             Inventory inv = Bukkit.createInventory(null, 9, "Upgrade your bank storage?");
                         	int invLvl = (int) DatabaseAPI.getInstance().getData(EnumData.INVENTORY_LEVEL, player.getUniqueId());
-                			int num = (int) invLvl * 100;
+                			int num = invLvl * 100;
                             ItemStack accept = new ItemStack(Material.WOOL, 1, DyeColor.LIME.getData());
                             ItemMeta acceptMeta = accept.getItemMeta();
                             acceptMeta.setDisplayName(ChatColor.GREEN.toString() +"Accept");
@@ -390,7 +390,7 @@ public class BankListener implements Listener {
         }else if(e.getInventory().getTitle().contains("Upgrade your bank storage")){
         	e.setCancelled(true);
         	int invLvl = (int) DatabaseAPI.getInstance().getData(EnumData.INVENTORY_LEVEL, player.getUniqueId());
-			int num = (int) BankMechanics.getPrice(invLvl);
+			int num = BankMechanics.getPrice(invLvl);
 			//TODO PRICE OF UPGRADE ^ 
         	int slot = e.getRawSlot();
         	if(slot ==3){
